requirements.txt
# Flask
Flask==1.1.1
Flask-WTF==0.14.3
Flask-Login==0.5.0
email-validator==1.1.0

# Database
flask-mongoengine==0.9.5

# Oauth
Authomatic==1.0.0

# Environment variables managment
python-dotenv==0.13.0
